+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# JIRA Daily candidates
+# Atlassian helpers
+
+## 0.0.6
+
+- Converted to Chrome extension
 
 ## 0.0.5